Move self-search functions to selfsearch.c
